Update for Elm 0.17. Bump version 5.0.1
Fix tests for the less flexible date parsing that Firefox does compared to Chrome.
Dates have been modified to an ISO format with offset, or are fudged to that with a TestUtil.fudgeDate now.
Test now should run on Firefox and Chrome.
Add test.cmd to launch testing in console on windows conveniently.
Update example programs.
OldDateFromFields removed along with the local copy of Benchmark from 0.16.