Add Greek and handle AM/PM in multiple languages
This is in one commit because:

* It seemed like a pre-requisite to add time properly in greek.
* Needed a language to test AM/PM

Some of the changes are due to elm-format running in previously
unformatted files.

Not fixed in this commit:
Uncovered the case for having multiple month suffixes
depending on context (see comments in Greek config for example).
This also applies to Russian. Requires quite a bit of
thinking to solve this properly.